+++ src/Auctionify.API/Controllers/SubscriptionsController.cs
@@ -1,4 +1,5 @@
 ﻿using Auctionify.Application.Features.Subscriptions.Commands;
+using Auctionify.Application.Features.Subscriptions.Commands.CreateProSubscription;
 using Auctionify.Application.Features.Subscriptions.Commands.DeleteProSubscription;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
